admin_panel/pyproject.toml
[project]
name = "movies-admin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.2",
    "django-debug-toolbar>=6.2.0",
    "django-split-settings>=1.3.2",
    "gunicorn>=25.1.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.2.0",
    "django-stubs-ext>=5.2.9",
    "django-types>=0.23.0",
    "isort>=8.0.0",
    "ruff>=0.15.2",
    "ty>=0.0.18",
]

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py314"
exclude = [
    "manage.py",
    "migrations",
    "config",
    ".venv"
]

[tool.ruff.lint]
preview = true
select = ["E", "F", "B", "COM", "N", "Q", "LOG", "G", "RET", "I", "ANN"]
ignore = ["E203", "I001"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
use_parentheses = true
line_length = 88
skip = [
    "manage.py",
    "migrations",
    ".venv"
]


[tool.ty.src]
exclude = [
    "manage.py",
    "migrations",
    ".venv"
]
